pyproject.toml
[project]
name = "digital-squad"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=7.5.2",
    "groq>=0.19.0",
    "loguru>=0.7.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "wikipedia-api>=0.8.1",
]
